docs: address some Next complaints
