+++ pnpm-workspace.yaml
@@ -15,9 +15,9 @@ catalogs:
   ark:
     '@ark-ui/react': ^5.37.2
   pandacss:
-    '@pandacss/dev': ^1.11.5
-    '@pandacss/preset-panda': ^1.11.5
-    '@pandacss/types': ^1.11.5
+    '@pandacss/dev': ^1.12.0
+    '@pandacss/preset-panda': ^1.12.0
+    '@pandacss/types': ^1.12.0
   jsx:
     '@carbon/icons-react': ^11.85.0
     globby: ^16.2.2
